requirements.txt
Flask
flask-cors
psycopg2-binary
gunicorn
bcrypt
requests